roll executor into worker
