+++ snuba/cli/subscriptions.py
@@ -14,7 +14,6 @@ from snuba.environment import setup_logging, setup_sentry
 from snuba.redis import redis_client
 from snuba.subscriptions.consumer import TickConsumer
 from snuba.subscriptions.data import PartitionId
-from snuba.subscriptions.executor import SubscriptionExecutor
 from snuba.subscriptions.scheduler import SubscriptionScheduler
 from snuba.subscriptions.store import RedisSubscriptionDataStore
 from snuba.subscriptions.worker import (
@@ -182,7 +181,8 @@ def subscriptions(
                 else Topic(loader.get_default_topic_spec().topic_name)
             ),
             SubscriptionWorker(
-                SubscriptionExecutor(dataset, executor, metrics),
+                dataset,
+                executor,
                 {
                     index: SubscriptionScheduler(
                         RedisSubscriptionDataStore(
@@ -200,6 +200,7 @@ def subscriptions(
                 },
                 producer,
                 Topic(result_topic),
+                metrics,
             ),
             max_batch_size,
             max_batch_time_ms,
