+++ snuba/subscriptions/worker.py
@@ -2,19 +2,24 @@ from __future__ import annotations
 
 import itertools
 import json
-from concurrent.futures import Future, as_completed
+import time
+from concurrent.futures import Future, ThreadPoolExecutor, as_completed
 from typing import Mapping, NamedTuple, Optional, Sequence
 
+from snuba.datasets.dataset import Dataset
 from snuba.reader import Result
 from snuba.subscriptions.consumer import Tick
 from snuba.subscriptions.data import Subscription
-from snuba.subscriptions.executor import SubscriptionExecutor
 from snuba.subscriptions.scheduler import ScheduledTask, Scheduler
 from snuba.utils.codecs import Codec
+from snuba.utils.metrics.backends.abstract import MetricsBackend
+from snuba.utils.metrics.gauge import Gauge
+from snuba.utils.metrics.timer import Timer
 from snuba.utils.streams.batching import AbstractBatchWorker
 from snuba.utils.streams.kafka import KafkaPayload
 from snuba.utils.streams.producer import Producer
 from snuba.utils.streams.types import Message, Topic
+from snuba.web.query import parse_and_run_query
 
 
 class SubscriptionResultFuture(NamedTuple):
@@ -30,15 +35,43 @@ class SubscriptionResult(NamedTuple):
 class SubscriptionWorker(AbstractBatchWorker[Tick, Sequence[SubscriptionResultFuture]]):
     def __init__(
         self,
-        executor: SubscriptionExecutor,
+        dataset: Dataset,
+        executor: ThreadPoolExecutor,
         schedulers: Mapping[int, Scheduler[Subscription]],
         producer: Producer[SubscriptionResult],
         topic: Topic,
+        metrics: MetricsBackend,
     ) -> None:
+        self.__dataset = dataset
         self.__executor = executor
         self.__schedulers = schedulers
         self.__producer = producer
         self.__topic = topic
+        self.__metrics = metrics
+
+        self.__concurrent_gauge = Gauge(self.__metrics, "executor.concurrent")
+
+    def __execute(self, task: ScheduledTask[Subscription], tick: Tick) -> Result:
+        # Measure the amount of time that took between this task being
+        # scheduled and it beginning to execute.
+        self.__metrics.timing(
+            "executor.latency", (time.time() - task.timestamp.timestamp()) * 1000
+        )
+
+        # XXX: The ``query`` name is taken from the web views so that all query
+        # performance metrics are reported to the same spot, regardless of
+        # execution environment.
+        timer = Timer("query")
+
+        request = task.task.data.build_request(
+            self.__dataset, task.timestamp, tick.offsets.upper, timer,
+        )
+
+        with self.__concurrent_gauge:
+            # XXX: The ``extra`` is discarded from ``QueryResult`` since it is
+            # not particularly useful in this context and duplicates data that
+            # is already being published to the query log.
+            return parse_and_run_query(self.__dataset, request, timer).result
 
     def process_message(
         self, message: Message[Tick]
@@ -54,7 +87,9 @@ class SubscriptionWorker(AbstractBatchWorker[Tick, Sequence[SubscriptionResultFu
         # rebalancing) and cause the entire batch to be have to be replayed.
         tick = message.payload
         return [
-            SubscriptionResultFuture(task, self.__executor.execute(task, tick))
+            SubscriptionResultFuture(
+                task, self.__executor.submit(self.__execute, task, tick)
+            )
             for task in self.__schedulers[message.partition.index].find(tick.timestamps)
         ]
